1주차 Modify 구조변경, Handler변경, 싱글톤 변수관리  관련
