+++ app/src/main/java/com/work/restaurant/view/fragment/search/SearchFragment.kt
@@ -1,22 +1,16 @@
-package com.work.restaurant.view.fragment
+package com.work.restaurant.view.fragment.search
 
 import android.content.Context
 import android.os.Bundle
-import android.os.Handler
 import android.util.Log
 import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
-import androidx.core.os.postDelayed
 import androidx.fragment.app.Fragment
 import com.work.restaurant.R
-import kotlinx.android.synthetic.main.loading_fragment.*
-import java.util.*
 
-class LoadingFragment : Fragment() {
-
-
-    private val fragmentName = "LoadingFragment"
+class SearchFragment : Fragment() {
+    private val fragmentName = "SearchFragment"
 
     override fun onAttach(context: Context) {
         Log.d(fragmentName, "onAttach")
@@ -31,34 +25,12 @@ class LoadingFragment : Fragment() {
 
     }
 
-
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
     ): View? {
-        return inflater.inflate(R.layout.loading_fragment, container, false)
-    }
-
-
-    override fun onActivityCreated(savedInstanceState: Bundle?) {
-        Log.d(fragmentName, "onActivityCreated")
-        super.onActivityCreated(savedInstanceState)
-
-
-        val random = Random().nextInt(resources.getStringArray(R.array.load_string).size)
-        val loadingTextArrayList = resources.getStringArray(R.array.load_string)
-
-        Log.d(fragmentName, loadingTextArrayList[random])
-
-        loading_tv.text = loadingTextArrayList[random]
-
-
-        Handler().postDelayed(3000L) {
-            requireFragmentManager().beginTransaction().remove(this).commit()
-        }
-
-
+        return inflater.inflate(R.layout.search_fragment, container, false)
     }
 
     override fun onStart() {
@@ -96,6 +68,4 @@ class LoadingFragment : Fragment() {
         super.onDetach()
     }
 
-
 }
-
