+++ app/src/main/res/drawable/bottom_line_2.xml
@@ -2,13 +2,14 @@
 <layer-list xmlns:android="http://schemas.android.com/apk/res/android">
     <item>
         <shape android:shape="rectangle">
-            <solid android:color="#FFFFFF" />
+            <solid android:color="@color/colorGray" />
         </shape>
     </item>
     <item android:bottom="2dp">
         <shape android:shape="rectangle">
-            <solid android:color="@color/colorGray" />
+            <solid android:color="#FFFFFF" />
         </shape>
     </item>
 
 </layer-list>
+
