+++ app/src/main/res/layout/activity_main.xml
@@ -5,7 +5,7 @@
     android:layout_width="match_parent"
     android:layout_height="match_parent"
     android:orientation="vertical"
-    tools:context=".view.activity.MainActivity">
+    tools:context=".view.MainActivity">
 
     <!--    <FrameLayout-->
     <!--        android:id="@+id/main_container"-->
