+++ app/src/main/res/layout/community_fragment.xml
@@ -2,8 +2,9 @@
 <FrameLayout xmlns:android="http://schemas.android.com/apk/res/android"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    android:orientation="vertical">
-
+    android:orientation="vertical"
+    android:id="@+id/community_container"
+    >
 
     <TextView
         android:layout_width="match_parent"
