+++ app/src/main/res/layout/loading_fragment.xml
@@ -11,7 +11,7 @@
         android:id="@+id/loading_tv"
         android:layout_width="match_parent"
         android:layout_height="wrap_content"
-        android:background="@drawable/bottom_line"
+        android:background="@drawable/bottom_line_1"
         android:gravity="center_horizontal"
         android:textSize="40dp"
         android:textStyle="bold">
@@ -25,7 +25,7 @@
         android:textSize="40dp"
         android:gravity="center_horizontal"
         android:textStyle="bold"
-        android:background="@drawable/bottom_line"
+        android:background="@drawable/bottom_line_1"
         >
 
     </TextView>
