+++ app/src/main/res/values/colors.xml
@@ -7,4 +7,5 @@
     <color name="colorWhite">#FFFFFF</color>
     <color name="colorBlue">#4180b9</color>
     <color name="colorBlack">#000000</color>
+    <color name="colorMint">#1ABC9C</color>
 </resources>
