+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,6 @@
 <resources>
     <string name="app_name">ExerciseRestaurant</string>
-
+    <string name="app_version">v.1.1.1</string>
+    <string name="app_update">Update 2019.10.18</string>
 
 </resources>
